pyproject.toml
[tool.poetry]
name = "py-eventbus-pg"
version = "0.1.0"
description = ""
authors = ["sampot <sam@i-me.tw>"]

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.25.0"
python-dotenv = "^0.19.2"
cuid = "^0.3"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = {version = "^22.1.0", allow-prereleases = true}
pytest-asyncio = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
